% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnomics_dwn.R
\name{dbnomics_dwn}
\alias{dbnomics_dwn}
\title{A wrapper around `rdbnomics::rdb()`}
\usage{
dbnomics_dwn(...)
}
\description{
This wrapper transforms the variable "original_period" to a numeric variable named `year` when
data are annual. Output is also transformed into a tibble.
}
