% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_colors.R
\name{rse_colors}
\alias{rse_colors}
\title{Function to extract RSE colors as hex codes}
\usage{
rse_colors(...)
}
\arguments{
\item{...}{Character names of RSE colors: mint , blue, pink, light green
light blue, pale green, pale blue, gray}
}
\description{
Function to extract RSE colors as hex codes
}
