% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_pal.R
\name{rse_pal}
\alias{rse_pal}
\title{Return a function to interpolate RSE palettes}
\usage{
rse_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in res_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return a function to interpolate RSE palettes
}
