% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_theme.R
\name{rse_theme}
\alias{rse_theme}
\title{A standardized ggplot theme for RSE's plots in reports and presentations}
\usage{
rse_theme(
  title_size = 22,
  title_family = "Verdana",
  text_family = "Verdana",
  legend_family = "Verdana",
  title_color = "seagreen4",
  text_color = "gray20",
  legend_color = "gray20",
  legend_pos = "bottom",
  axis_face = "plain",
  text_rel = 1,
  strip_rel = 1,
  axis_rel = 1,
  legend_rel = 0.8,
  caption_rel = 0.65
)
}
\arguments{
\item{title_size}{A numeric value. The size of the title in points. Default is `22`. This
paramters determines - indirectly - also the sie of the subtitle.}

\item{title_color, text_color, legend_color}{The text color of title, regular text and legend. The `title_color`
paramters also sets the caption color.}

\item{legend_pos}{the position of legends ("none", "left", "right", "bottom", "top",
or two-element numeric vector)}

\item{axis_face}{Font face ("plain", "italic", "bold", "bold.italic") of the axis text. Default is "plain".}

\item{text_rel, strip_rel, axis_rel, legend_rel, caption_rel}{Relative size of various elements of text.}

\item{title_familiy, text_family, legend_familiy}{The font type for title, text, and legend. Default
is set to "Verdana". The package automatically loads all fonts installed in the system (windows only).}
}
\description{
This function allows you to add a standardized theme to your ggplot graphics
}
\examples{
\dontrun{
ggplot(mtcars, aes(x = mpg, y = disp, color = hp)) +
geom_point() +
rse_theme()
}
}
