% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_colour_rse}
\alias{scale_colour_rse}
\alias{scale_fill_rse}
\title{Color scale constructor for RSE colors}
\usage{
scale_colour_rse(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_rse(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in rse_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for RSE colors

Fill scale constructor for rse colors
}
