% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsgm-package.R
\docType{package}
\name{utilsgm-package}
\alias{utilsgm}
\alias{utilsgm-package}
\title{utilsgm: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giuliomela/utilsgm}
  \item Report bugs at \url{https://github.com/giuliomela/utilsgm/issues}
}

}
\keyword{internal}
