% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_colour_rse}
\alias{scale_colour_rse}
\alias{scale_fill_rse}
\title{Create color scales for ggplots}
\usage{
scale_colour_rse(discrete = TRUE)

scale_fill_rse(discrete = TRUE)
}
\description{
These functions create a color/fill scale (with the persoalized RSE palette)
}
